Created a separate threadStarter() method to populate the processingQueue without having to run the comicInformationGrabber all the time. This is helpful for when save states are implemented, so we don't have to grab the comic information everytime.
Recoded the implementation of integers and doubles in the code. Instead, each variable is of type Object so it is easier to work with and modify according to needs.

Added TODO for a review of Big O Notation. Might be better to work access URLs backwards to add them to a LinkedList in O(n) time instead of whatever the time notation is currently.